Changed functionality in stream method.
The Element on which the stream method is called is also completely
streamed.  This is more intuitive,
the stream method is called when the user is finished with this element.
The element must be also completely streamed.

The stream method doesn't return the Element that it calls, because it
is already streamed and should not be used anymore!